♻️ (server): Move usecase defintion file
