+++ apps/server/src/bookmarks/applications/usecases/delete-bookmark-with-auth/delete-bookmark-with-auth.usecase.ts
@@ -1,6 +1,6 @@
 import { InjectRepository } from '@nestjs/typeorm';
 import { BookmarkUsersRepository } from '@readable/bookmarks/infrastructures/typeorm/repositories/bookmarkUsers.repository';
-import { Usecase } from '@readable/common/usecase';
+import { Usecase } from '@readable/common/applications/usecase';
 import { User } from '@readable/users/domain/models/user.model';
 import { DeleteBookmarkWithAuthInput } from './delete-bookmark-with-auth.input';
 import {
