+++ apps/server/src/bookmarks/applications/usecases/get-my-bookmarks/get-my-bookmarks.usecase.ts
@@ -2,7 +2,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Bookmark } from '@readable/bookmarks/infrastructures/typeorm/entities/bookmark.entity';
 import { BookmarksRepository } from '@readable/bookmarks/infrastructures/typeorm/repositories/bookmarks.repository';
 import { BookmarkUsersRepository } from '@readable/bookmarks/infrastructures/typeorm/repositories/bookmarkUsers.repository';
-import { Usecase } from '@readable/common/usecase';
+import { Usecase } from '@readable/common/applications/usecase';
 import { User } from '@readable/users/domain/models/user.model';
 import { In } from 'typeorm';
