+++ apps/server/src/pagination/paginationBookmarks/applications/usecases/get-pagination-bookmarks/get-pagination-bookmarks.usecase.ts
@@ -1,6 +1,6 @@
 import { InjectRepository } from '@nestjs/typeorm';
 import { BookmarksRepository } from '@readable/bookmarks/infrastructures/typeorm/repositories/bookmarks.repository';
-import { Usecase } from '@readable/common/usecase';
+import { Usecase } from '@readable/common/applications/usecase';
 import { PaginationBookmarkBRFOs } from '@readable/pagination/paginationBookmarks/domain/models/paginationBookmarks.type';
 import { User } from '@readable/users/domain/models/user.model';
 import { GetPaginationBookmarksInput } from './get-pagination-bookmarks.input';
